backend/requirements.txt
Adafruit-Blinka
adafruit-circuitpython-fingerprint
adafruit-circuitpython-lis3dh
Adafruit-PlatformDetect
Adafruit-PureIO
autopep8
certifi
charset-normalizer
click
dj-database-url
Django
django-heroku
dlib
face-recognition
face-recognition-models
gunicorn
idna
numpy
opencv-python
pep8
Pillow
psycopg2
pycodestyle
pyftdi
PyJWT
pyserial
pytz
pyusb
requests
sqlparse
twilio
urllib3
whitenoise
